update week6 homework

+++ src/main/java/week6/homework/tables_ddl_06/tables_ddl.sql
@@ -9,8 +9,8 @@ create table java_course.SHOP_ORDER
     IS_PAID tinyint(1) default 0 not null,
     IS_CANCELED tinyint(1) default 0 not null,
     IS_DELIVERED tinyint(1) default 0 not null,
-    CREATE_TIME int not null,
-    UPDATE_TIME int not null
+    CREATE_TIME bigint not null,
+    UPDATE_TIME bigint not null
 );
 
 create table java_course.SHOP_ORDER_DETAIL
@@ -18,8 +18,8 @@ create table java_course.SHOP_ORDER_DETAIL
     ORDER_ID varchar(16) not null,
     PRODUCT_ID varchar(16) not null,
     ORDER_QUANTITY int not null,
-    CREATE_TIME int not null,
-    UPDATE_TIME int not null
+    CREATE_TIME bigint not null,
+    UPDATE_TIME bigint not null
 );
 
 create table java_course.SHOP_PRODUCT
@@ -31,8 +31,8 @@ create table java_course.SHOP_PRODUCT
     PRODUCT_PRICE decimal(14,4) not null,
     WAREHOUSE_ID varchar(16) not null,
     WAREHOUSE_QUANTITY int not null,
-    CREATE_TIME int not null,
-    UPDATE_TIME int not null
+    CREATE_TIME bigint not null,
+    UPDATE_TIME bigint not null
 );
 
 create table java_course.SHOP_USER
@@ -43,8 +43,8 @@ create table java_course.SHOP_USER
     EMAIL varchar(64) null,
     PHONE varchar(16) null,
     LAST_LOGIN int null,
-    CREATE_TIME int not null,
-    UPDATE_TIME int not null
+    CREATE_TIME bigint not null,
+    UPDATE_TIME bigint not null
 );
 
 create table java_course.SHOP_WAREHOUSE
@@ -53,6 +53,7 @@ create table java_course.SHOP_WAREHOUSE
         primary key,
     WAREHOUSE_ADDRESS varchar(128) not null,
     MANAGER varchar(32) null,
-    CREATE_TIME int not null,
-    UPDATE_TIME int not null
+    CREATE_TIME bigint not null,
+    UPDATE_TIME bigint not null
 );
+
